I24 serial test fixes

+++ src/mx_bluesky/I24/serial/fixed_target/i24ssx_Chip_StartUp_py3v1.py
@@ -64,9 +64,9 @@ def scrape_parameter_file(param_path: Path | str = PARAM_FILE_PATH_FT):
         if "dcdetdist" in entry[0].lower():
             dcdetdist = float(entry[1])
         if "prepumpexptime" in entry[0].lower():
-            prepumpexptime = int(entry[1])
+            prepumpexptime = float(entry[1])
         if "pumpdelay" in entry[0].lower():
-            pumpdelay = int(entry[1])
+            pumpdelay = float(entry[1])
         if "det_type" in entry[0].lower():
             det_type = entry[1]
     return (
